Filter out null roles in ldap settings
When setting a role as additional default group and deleting it while
it is still in the configuration, the backend will fail to resolve the
role name and return a `null` instead. This null element is stored in
MongoDB (and also returned from the API), which causes issues in some
circumstances.

Fixes #7431